+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 #### Bugfixes
 
 - Fixed a crash due to double borrow when handling pre/post run hooks
+- Fixes a panic that can occur when large `Duration`s are passed to `Timer::from_duration`.
 
 #### Additions